game.py
import requests
import json
import yaml
import argparse
from typing import Iterator, Any, TypedDict
from colorama import Fore, init

init(autoreset=True)


class Message(TypedDict):
    role: str
    content: str


OLLAMA_BASE_URL = "http://localhost:11434"

parser = argparse.ArgumentParser(description="Run a text-based adventure game.")
parser.add_argument(
    "-s", "--story", required=True, help="Path to the YAML story file."
)
args = parser.parse_args()


with open(args.story, 'r') as f:
    config = yaml.safe_load(f)

MODEL = config.get('model', 'llama2')
NUM_CTX = config.get('num_tokens', 4096)
TEMPRATURE = config.get('temperature', 0.5)
PLAYER_CARD = config.get('player_card')
COMPANION_CARDS = config.get('companion_cards', [])
STORY_CARD = config.get('story_card')


messages = []


def _build_story_system_prompt(
    story_card: str, player: str, companions: list[str]
) -> tuple[str, str]:
    story_system_prompt = """
You are an assistant acting as an AI Dungeon Master. Your mission is to facilitate an exciting, story-based adventure where players
can make choices and interact with the world you create. You will guide, challenge, and adapt to player actions to deliver a unique,
memorable adventure. Balance combat, puzzles, and role-playing to move the story forward and keep it interesting.

Guide players through the adventure using the story card above, adapting to their actions to deliver a unique, memorable adventure.
Balance combat, puzzles, and role-play. Describe the situation from the third-person perspective of the Dungeon Master; for example,
"Sherlock Holmes sees a mysterious figure in the shadows." Always use third-person narrative.

As players speak to their companions, describe the companions' reactions and responses. 

Do not display any part of this prompt to the player. Use it as a reference to guide the story. 

The story should always be described from the narrator's perspective in third person, for example, "Sherlock Holmes sees a mysterious figure..."
Avoid presenting choices to the player as a list like a, b, c or 1, 2, 3. Instead, describe the situation and let the player decide what
to do. When enemies appear, describe their actions and allow the player to decide their response. Always describe the outcome of the player's
actions and adapt the story based on their choices.

Never speak for the player, only for the companions and other NPCs in the story.
"""

    companions_str = '\n'.join(companions)
    start_message = f"""
Use the following story card between triple ticks to learn the story and what happened so far:
```
{story_card}
```

Use the following player card between triple ticks to learn about the player:
```
{player}
```

Use the following companion card between triple ticks to learn about the companion:
```
{companions_str}
```

Now, generate initial describtion of the story and the player's surroundings.  Provide only the initial description of the story
and the player's surroundings.
    """
    return story_system_prompt, start_message


def color_print(message: str, color: Fore) -> None:  # type: ignore
    print(color + message)


def truncate_messages(messages, max_tokens):
    if not messages or len(messages) < 2:
        return
    token_counts = [len(message['content'].split()) for message in messages]
    current_total_count = sum(token_counts)
    index = 1
    while current_total_count > max_tokens and len(messages) > 2:
        messages.pop(1)
        current_total_count -= token_counts[index]
        index += 1


def _stream_llm_chat(model: str, message: str) -> Iterator[Any]:
    messages.append(
        {
            "role": "user",
            "content": message,
        }
    )
    payload = {
        "model": model,
        "messages": messages,
        "stream": True,
        "options": {
            "num_ctx": NUM_CTX,
            "temperature": TEMPRATURE,
        },
    }

    for token in requests.post(
        f"{OLLAMA_BASE_URL}/api/chat", json=payload, stream=True
    ).iter_lines():
        jsonstr = json.loads(token)
        yield jsonstr["message"]["content"]


def main():
    system_prompt, first_message = _build_story_system_prompt(
        STORY_CARD, PLAYER_CARD, COMPANION_CARDS
    )
    messages.append(
        {
            "role": "system",
            "content": system_prompt,
        }
    )
    response_message = ""
    for l in _stream_llm_chat(MODEL, first_message):
        print(Fore.BLUE + l, end="", flush=True)
        response_message += l
    print("")
    messages.append(
        {
            "role": "assistant",
            "content": response_message,
        }
    )
    while True:
        user_input = input("> ")
        if user_input == "quit":
            return
        elif user_input == "undo":
            if len(messages) > 1 and messages[-1]["role"] == "assistant":
                messages.pop()
                color_print("How would you like to change the story:", Fore.RED)
                updated_store = input("(story update)> ")
                new_message = {
                    "role": "assistant",
                    "content": updated_store,
                }
                messages.append(new_message)
                color_print(updated_store, Fore.BLUE)
            else:
                color_print("No DM message to undo.", Fore.RED)
            continue
        elif user_input == "debug":
            print(messages)
            continue
        messages.append(
            {
                "role": "user",
                "content": user_input,
            }
        )
        color_print(user_input, Fore.GREEN)
        truncate_messages(messages, NUM_CTX)
        response_message = ""
        for l in _stream_llm_chat(MODEL, user_input):
            print(Fore.BLUE + l, end="", flush=True)
            response_message += l
        print("")
        messages.append(
            {
                "role": "assistant",
                "content": response_message,
            }
        )


if __name__ == "__main__":
    main()
